[MODIFY-27/01-11:00]PERSONAL DATA service new method, exception handling

+++ src/main/java/by/vorakh/alex/spring_company/service/PersonalDataService.java
@@ -3,12 +3,15 @@ package by.vorakh.alex.spring_company.service;
 import java.util.ArrayList;
 import java.util.List;
 
+import javax.persistence.EntityExistsException;
 import javax.transaction.Transactional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import by.vorakh.alex.spring_company.converter.PersonalDataOutsourceToPersonalDataConverter;
 import by.vorakh.alex.spring_company.converter.PersonalDataToPersonalDataViewModelConverter;
+import by.vorakh.alex.spring_company.model.outsource.PersonalDataOutsource;
 import by.vorakh.alex.spring_company.model.payload.PersonalDataPayload;
 import by.vorakh.alex.spring_company.model.view_model.IdViewModel;
 import by.vorakh.alex.spring_company.model.view_model.PersonalDataViewModel;
@@ -25,6 +28,8 @@ public class PersonalDataService implements ServiceInterface<PersonalDataViewMod
     private EmployeeDAO employeeDAO;
     @Autowired
     private PersonalDataToPersonalDataViewModelConverter convertor;
+    @Autowired
+    private PersonalDataOutsourceToPersonalDataConverter extermalSourceConvertor;
     
     @Override
     public List<PersonalDataViewModel> getAll() {
@@ -41,12 +46,67 @@ public class PersonalDataService implements ServiceInterface<PersonalDataViewMod
 	return convertor.convert(personalDataDAO.getById(id));
     }
     
+    @SuppressWarnings("finally")
     @Override
     @Transactional
     public IdViewModel create(PersonalDataPayload newPayload) {
-	return new IdViewModel()
-		.setId(personalDataDAO
-			.create(new PersonalData(newPayload.getFirstName(), newPayload.getLastName())));
+	int createdID = -1;
+	
+	PersonalData newPersonalData = new PersonalData(newPayload.getFirstName(), newPayload.getLastName());
+	try {
+	    
+	    createdID= personalDataDAO
+			.create(newPersonalData);
+	} catch (EntityExistsException e) {
+	    throw new ServiceException("The \"" + newPersonalData.getFirstName() + 
+    		    " " + newPersonalData.getLastName() +  
+		    "\" cannot be created, because to exist in database.", e);
+	} catch (IllegalArgumentException ex) {
+	    throw new ServiceException("The PERSONAL_DATA cannot be created, because " + 
+		    newPersonalData.toString() +  " is not a PERSONAL_DATA object.", ex);
+	}  catch (javax.persistence.TransactionRequiredException exc) { 
+	    throw new ServiceException("The \"" + newPersonalData.getFirstName() + 
+    		    " " + newPersonalData.getLastName() +  
+		    "\" cannot be created, NO Transaction.", exc);
+	} catch (javax.persistence.PersistenceException ex1) { 
+	    throw new ServiceException("The \"" + newPersonalData.getFirstName() + 
+    		    " " + newPersonalData.getLastName() +  
+		    "\" cannot be created, the database is not updated.", ex1);
+	} finally {
+	    return new IdViewModel(createdID);
+	}
+    }
+    
+    @Transactional
+    public PersonalDataViewModel createAndGet(PersonalDataOutsource extemalSource) {
+	PersonalData newPersonalData = extermalSourceConvertor.convert(extemalSource);
+	return createAndGet(newPersonalData);
+    }
+
+    @SuppressWarnings("finally")
+    @Transactional
+    public PersonalDataViewModel createAndGet(PersonalData newPersonalData) {
+	PersonalData returnedPersonalData = null;
+    	try { 
+    	    returnedPersonalData = personalDataDAO.createAndGet(newPersonalData);
+    	} catch (EntityExistsException e) {
+    	    throw new ServiceException("The \"" + newPersonalData.getFirstName() + 
+    		    " " + newPersonalData.getLastName() + 
+    		    "\" cannot be created, because to exist in database.", e);
+	} catch (IllegalArgumentException ex) {
+    	    throw new ServiceException("The PERSONAL_DATA cannot be created, because " + 
+    		    newPersonalData.toString() +  " is not a Skill object.", ex);
+    	} catch (javax.persistence.TransactionRequiredException exc) { 
+    	    throw new ServiceException("The \"" + newPersonalData.getFirstName() + 
+    		    " " + newPersonalData.getLastName() +  
+    		    "\" cannot be created, NO Transaction.", exc);
+    	} catch (javax.persistence.PersistenceException ex1) { 
+    	    throw new ServiceException("The \"" + newPersonalData.getFirstName() + 
+    		    " " + newPersonalData.getLastName() +  
+    		    "\" cannot be created, the database is not updated.", ex1);
+    	} finally {
+    	    return convertor.convert(returnedPersonalData);
+    	}
     }
     
     @Override
@@ -56,15 +116,37 @@ public class PersonalDataService implements ServiceInterface<PersonalDataViewMod
 	personalDataForEditing
 		.setFirstName(editedPayload.getFirstName())
 		.setLastName(editedPayload.getLastName());
-	personalDataDAO.update(personalDataForEditing);
+	try {
+	    personalDataDAO.update(personalDataForEditing);
+	} catch (IllegalArgumentException ex) {
+	    throw new ServiceException("The PERSONAL_DATA cannot be updated, because " + 
+		    personalDataForEditing.toString() +  " is not a Skill object.", ex);
+	} catch (javax.persistence.TransactionRequiredException exc) { 
+	    throw new ServiceException("The \"" + personalDataForEditing.getFirstName() + 
+    		    " " + personalDataForEditing.getLastName() +  
+		    "\" cannot be updated, NO Transaction.", exc);
+	}
     }
 
     @Override
     @Transactional
     public void delete(int id) {
 	PersonalData deletedPersonalData = personalDataDAO.getById(id);
+	if (deletedPersonalData == null) {
+	    throw new ServiceException("The PERSONAL_DATA with \'" + id + 
+		    "\' ID cannot be deleted, because to no exist in database");
+	}
 	employeeDAO.delete(deletedPersonalData);
-	personalDataDAO.delete(deletedPersonalData);
+	try {
+	    personalDataDAO.delete(deletedPersonalData);
+	} catch (IllegalArgumentException ex) {
+	    throw new ServiceException("The PERSONAL_DATA cannot be deleted, because " + 
+		    deletedPersonalData.toString() +  " is not a Skill object.", ex);
+	} catch (javax.persistence.TransactionRequiredException exc) { 
+	    throw new ServiceException("The \"" + deletedPersonalData.getFirstName() + 
+    		    " " + deletedPersonalData.getLastName() +  
+		    "\" cannot be deleted, NO Transaction.", exc);
+	}
     }
 
 }
